Window, first iso2d
Just setting up the initial isometric 2d rendering system.

We're going to start with a single layer. I'm going to extend this to render adjacent chunks of tiles.

We'll probably just end up with a little random loop that goes to a position and decides whether or not to select  a chunk. Idk, it's 2AM